+++ server-app/Database/Entities/studentEntity.ts
@@ -1,0 +1,8 @@
+export interface studentEntity{
+    id: number,
+    name: string,
+    fname: string,
+    email: string,
+    phoneNumber: string
+}
+
