+++ server-app/Database/Entities/teacherEntity.ts
@@ -1,11 +1,3 @@
-export interface studentEntity{
-    id: number,
-    name: string,
-    fname: string,
-    email: string,
-    phoneNumber: string
-}
-
 export interface teacherEntity{
     id: number,
     name: string,
